src/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["mateusmbg <mateus.gois@sou.inteli.edu.br>"]
readme = "README.md"
packages = [{ include = "etl_bronze" }]

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.0"
seaborn = "^0.12"
matplotlib = "^3.7"
jupyter = "^1.0"
scikit-learn = "^1.3"
boto3 = "^1.28.0"
flask = "^3.0.3"
requests = "^2.32.3"
minio = "^7.2.10"
python-dotenv = "^1.0.1"
clickhouse-connect = "^0.8.6"
pyarrow = "^18.0.0"
pydantic-settings = "^2.6.1"
clickhouse-driver = "^0.2.9"
pydantic = "^2.9.2"
prefect = "^3.1.3"
flasgger = "^0.9.7.1"
streamlit-echarts = "^0.4.0"
psycopg2-binary = "^2.9.10"
numpy = "^2.1.3"
streamlit = "^1.40.2"
aws = "^0.2.5"
awscli = "^1.36.12"
moto = "^5.0.21"
plotly = "^5.24.1"
botocore = "^1.35.81"

[tool.poetry.scripts]
run_etl_pipeline = "etl_bronze:run_etl_pipeline"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
